requirements.txt
Pillow==9.2.0
matplotlib==3.6.3
numpy==1.24.0
streamlit==1.32.0
tensorflow==2.15.0